feat: implement `transformSources()` w/ domain model
Introduces a domain model for _Package_ and _Entry Point_. Serves as a foundation for a set of interesting ideas: what if... we could analyze an entry points `import` statements and dynamically create a minimal rollup configuration? What if... we further can detect what entry points depends on another entry point (within the same package) and we could create a build graph and automatically adjust the tsconfig?

BREAKING CHANGES:
Introduces a domain model for _Package_ and _Entry Point_ (as defined in Angular Package Format Specification Glossary). Albeit the refactoring was undertaken with care and integration tests were not changed, it may cause undesired behaviour.